requirements.txt
opencv-python~=4.11.0.86
numpy~=1.26.4
matplotlib~=3.8.4
pillow~=11.1.0
pygrabber
torch~=2.1.0+cu121
scikit-image~=0.25.2
torchvision~=0.16.0+cu121
PyYAML~=6.0.2